feat (user) : user 도메인 vo 내부 클래스로 이동
